refactor: cleanup and correct interfaces
Interface and implementation function declaration differs (https://github.com/code-423n4/2021-04-maple-findings/issues/98)

- Corrected non-matching interface names
- Explicit uint256 usage
- Removed unused or non-existent interfaces
- Increase interface coverage
- Remove unused variables
- Set minimal explicit state-mutabilities
- Standardized "visibility | state-mutability" pattern for functions
- Fixed one different compiler version

+++ contracts/MplRewards.sol
@@ -151,7 +151,7 @@ contract MplRewards is Ownable {
         // This keeps the reward rate in the right range, preventing overflows due to
         // very high values of rewardRate in the earned and rewardsPerToken functions;
         // Reward + leftover must be less than 2^256 / 10^18 to avoid overflow.
-        uint balance = rewardsToken.balanceOf(address(this));
+        uint256 balance = rewardsToken.balanceOf(address(this));
         require(rewardRate <= balance.div(rewardsDuration), "R:REWARD_TOO_HIGH");
 
         lastUpdateTime = block.timestamp;
@@ -162,7 +162,7 @@ contract MplRewards is Ownable {
     /**
         @dev End rewards emission earlier. Only the contract Owner may call this.
     */
-    function updatePeriodFinish(uint timestamp) external onlyOwner {
+    function updatePeriodFinish(uint256 timestamp) external onlyOwner {
         _updateReward(address(0));
         periodFinish = timestamp;
     }
